requirements.txt
Flask==2.0.1
numpy==1.21.0
opencv-python-headless==4.5.3.56
Pillow==8.3.1